Kelas w12

+++ app/Http/Controllers/SupplierController.php
@@ -110,6 +110,9 @@ class SupplierController extends Controller
      */
     public function destroy(Supplier $supplier)
     {
+        // Week 12
+        $this->authorize('delete-permission', $supplier);
+        
         // Week 11 -> Hapus data
         $supplier->delete();
         return redirect()->route('suppliers.index')->with('status','data berhasil dihapus');
